requirements/base.txt
Django==2.0.7
psycopg2==2.7.4
Pillow==5.1.0
six==1.11.0
dj-database-url==0.5.0
whitenoise==3.3.1

# TODO: Remove at production. Fixes `ModuleNotFoundError` at Heroku Staging.
django-debug-toolbar==1.9.1
sqlparse==0.2.4
